リファクタリング。authやredis関連処理を関数化してsugo_hub.jsをスリムにした
