+++ lib/adaptors/.index.js.bud
@@ -1,0 +1,18 @@
+/**
+ * Bud file for index.js
+ */
+
+'use strict'
+
+const { indexJsBud } = require('ape-tmpl')
+
+module.exports = indexJsBud({
+  desc: 'Adaptors',
+  module: 'Socket io adaptors',
+  dirname: __dirname,
+  subclasses: []
+})
+
+if (!module.parent) {
+  require('coz').render(__filename)
+}
