+++ lib/sugo_hub.js
@@ -18,10 +18,7 @@ const co = require('co')
 const sgSocket = require('sg-socket')
 const sgServer = require('sg-server')
 const debug = require('debug')('sg:hub')
-const redis = require('redis').createClient
-const socketIORedis = require('socket.io-redis')
-const socketIOAuth = require('socketio-auth')
-const parseUrl = require('url').parse
+const { redisAdaptor, authAdaptor } = require('./adaptors')
 const newStorage = require('./helpers/new_storage')
 const ioInterceptor = require('./helpers/io_interceptor')
 
@@ -88,29 +85,13 @@ class SugoHub {
 
     if (storageConfig.redis) {
       let { url, host, port } = storageConfig.redis
-      let auth = ''
-      if (url) {
-        let parsed = parseUrl(url)
-        host = parsed.hostname || host
-        port = parsed.port || port
-        auth = parsed.auth ? parsed.auth.split(':')[ 1 ] : ''
-      }
-      let pub = redis(port, host, { auth_pass: auth })
-      let sub = redis(port, host, { auth_pass: auth, return_buffers: true })
-      let client = socketIORedis({ pubClient: pub, subClient: sub, key: `${prefix}:socket.io` })
-      io.adapter(client)
+      redisAdaptor(io, { url, host, port, prefix })
     }
 
     if (authenticate) {
-      let authenticateCall = (socket, data, callback) => {
-        Promise.resolve(authenticate(socket, data))
-          .then((result) => callback(null, result))
-          .catch((err) => callback(err))
-      }
-
-      socketIOAuth(actorIO, { authenticate: authenticateCall })
-      socketIOAuth(callerIO, { authenticate: authenticateCall })
-      socketIOAuth(observerIO, { authenticate: authenticateCall })
+      authAdaptor(actorIO, { authenticate })
+      authAdaptor(callerIO, { authenticate })
+      authAdaptor(observerIO, { authenticate })
     }
 
     let actorService = new ActorService(storage)
